requirements.txt
alembic==1.15.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
